RavenDB-18104 - GET /databases/*/subscriptions
